feat: separated style files for SideMenu and ContentContainer components of Algorithm page
