+++ src/components/Pages/Algorithms/ContentContainer/Section/index.tsx
